+++ mw/base.ts
@@ -31,8 +31,8 @@ declare interface IWikitextModel extends editor.ITextModel {
 }
 
 // 每次新增插件都需要修改这里
-const baseVersion = '2.18',
-	addons = ['colorPicker'];
+const baseVersion = '2.21',
+	addons = ['hover'];
 
 mw.loader.load(`${CDN}/${REPO_CDN}/mediawiki.min.css`, 'text/css');